action: add option to post the workflow URL
as a comment to a configurable comments API URL.
For instance when having very long running workflows called by the
action, it's hard for the user to see the actual pipeline and its
results.
For better tracebility and user expierence post the link to
the acquired workflow URL as a comment.

Typical setting would be to set it to
${{ github.event.pull_request.comments_url }} for pull requests.

As this may not be wanted by all users set the default to empty,
meaning no comments will be made